src/{email,ntpd,utils_cmd_getval}.c: Fix formatting errors.

+++ src/ntpd.c
@@ -613,7 +613,7 @@ static int ntpd_receive_response (int *res_items, int *res_size,
 		if (pkt_item_len < res_item_size)
 			pkt_padding = res_item_size - pkt_item_len;
 		DEBUG ("res_item_size = %i; pkt_padding = %i;",
-				res_item_size, pkt_padding);
+				res_item_size, (int) pkt_padding);
 
 		/* Extract the sequence number */
 		pkt_sequence = INFO_SEQ (res.auth_seq);
@@ -652,7 +652,7 @@ static int ntpd_receive_response (int *res_items, int *res_size,
 		 * We start by allocating some more memory.
 		 */
 		DEBUG ("realloc (%p, %i)", (void *) *res_data,
-				(items_num + pkt_item_num) * res_item_size);
+				(int) (items_num + pkt_item_num) * res_item_size);
 		items = realloc ((void *) *res_data,
 				(items_num + pkt_item_num) * res_item_size);
 		if (items == NULL)
